requirements.txt
aiofiles==0.8.0
aioprocessing==2.0.0
aiosqlite==0.17.0
APScheduler==3.9.1
asyncpg==0.26.0
dependency-injector==4.40.0
httptools==0.4.0
install==1.3.5
iso8601==1.0.2
loguru==0.6.0
multidict==6.0.2
orjson==3.8.0
prometheus-client==0.14.1
pydantic==1.10.0
pypika-tortoise==0.1.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
sanic==22.9.1
sanic-ext==22.9.1
sanic-routing==22.8.0
six==1.16.0
tortoise-orm==0.19.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
